% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsLossErrorMoment.R
\name{AbsLossErrorMoment}
\alias{AbsLossErrorMoment}
\title{\code{AbsLossErrorMoment} computes the conditional error moments}
\usage{
AbsLossErrorMoment(Raw, Pred, nu, sigma, w, p, Homoskedastic = TRUE)
}
\arguments{
\item{Raw}{Numeric vector with the raw values of the object variable for each statistical unit.}

\item{Pred}{Numeric vector with the predicted values for the object variable for each statistical
unit.}

\item{nu}{Numeric vector with the prediction error standard deviation for each statistical unit.}

\item{sigma}{Numeric vector with the observation error standard deviation for each statistical
unit.}

\item{w}{Numeric vector with the design weights for each statistical unit.}

\item{p}{Numeric vector with the probabilities of measurement error for each statistical unit.}

\item{Homoskedastic}{\code{TRUE} (default) or \code{FALSE} indicating whether to use a
homoskedastic or heteroskedastic model.}
}
\value{
Numeric vector with the values moments for each statistical unit.
}
\description{
\code{AbsLossErrorMoment} computes the conditional error moments
}
\examples{

AbsLossErrorMoment(10000, 12000, 2500, 5000, 1.9, 0.75)

}
