% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angle.R
\name{Angle}
\alias{Angle}
\title{C치lculo del (coseno al cuadrado del) 치ngulo entre dos vectores.}
\usage{
Angle(vector1, vector2)
}
\arguments{
\item{vector1}{\code{Vector} de tipo \code{numeric} con las componentes del primer vector.}

\item{vector2}{\code{Vector} de tipo \code{numeric} con las componentes del segundo vector.}
}
\value{
\code{vector} de tipo \code{numeric} de longitud 1 con el valor del coseno al cuadrado
del 치ngulo entre los vectores \code{vector1} y \code{vector2}.
}
\description{
\code{Angle} calcula el coseno cuadrado del 치ngulo entre dos vectores.
}
\examples{
Angle(c(1, 3, 5), c(0, -1, 1))

}
