% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KummerM.R
\name{KummerM}
\alias{KummerM}
\title{\code{KummerM} is a wrapper for the Kummer function in package fAsianOptions}
\usage{
KummerM(x, KummerLim = 50)
}
\arguments{
\item{x}{Numeric vector.}

\item{KummerLim}{Numeric value to control passing large numeric arguments to \code{kummerM}.}
}
\value{
Numeric vector with the value of the Kummer M function for each statistical unit.
}
\description{
\code{KummerM} is a wrapper for the Kummer function in package fAsianOptions
}
\examples{
KummerM(-23.2)
KummerM(-1200)

}
