% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logit.R
\name{Logit}
\alias{Logit}
\title{Función logit.}
\usage{
Logit(p, adjust = 1e-15)
}
\arguments{
\item{p}{\code{Vector} de tipo \code{numeric} con los datos para los que queremos calcular el
logit.}

\item{adjust}{\code{Vector} de tipo \code{numeric} de longitud 1. Por defecto toma el valor
\code{10^-15}.}
}
\value{
\code{Vector} de tipo \code{numeric} de longitud igual al vector de entrada \code{p} con
los logits de cada uno de los elementos en dicho vector.
}
\description{
\code{Logit} calcula el logit de cada una de las componentes del vector especificado en el
parámetro de entrada \code{p}. El parámetro de entrada \code{adjust} reduce en esta cantidad la
longitud del dominio [0,1] para evitar las singularidades en 0 y en 1.
}
\examples{
Logit(c(0.1, 0.3, 0.5, 0, -0.1, 0.1))

}
