% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinkUnitSc.R
\name{MinkUnitSc}
\alias{MinkUnitSc}
\title{\code{MinkUnitSc} computes the unit scores according to a Minkowskian function}
\usage{
MinkUnitSc(x, alpha = 1L, Weights = 1L)
}
\arguments{
\item{x}{Matrix with the item scores of each variable (columns) of each statistical unit (rows).}

\item{alpha}{Integer vector of length 1 with the degree of the Minkowskian function.}

\item{Weights}{Numeric vector with the weights of each item score.}
}
\value{
Numeric vector with the value of the Minkowskian function for each statistical unit.
}
\description{
\code{MinkUnitSc} computes the unit scores according to a Minkowskian function
}
\examples{

x <- cbind(c(0.434, 0.543, 1.445, 0.485), c(0.984, 1.032, 1.902, 0.840))
MinkUnitSc(x, 1, 1)

}
