% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeta.R
\name{Zeta}
\alias{Zeta}
\title{\code{zeta} computes the function \eqn{\zeta}}
\usage{
Zeta(p, nu, sigma, u)
}
\arguments{
\item{p}{Numeric vector with the probabilities of measurement error of each statistical unit.}

\item{nu}{Numeric vector with the prediction error standard deviation of each statistical unit.}

\item{sigma}{Numeric vector with the observation error standard deviation of each statistical
unit.}

\item{u}{Numeric vector with the predicted measurement error (observed value minus predicted
value) for each statistical unit.}
}
\value{
Numeric vector with the value of the zeta function for each statistical unit.
}
\description{
\code{zeta} computes the function \eqn{\zeta}
}
\examples{

Zeta(0.5, 130, 400, -20.3)

}
